pyproject.toml
[project]
name = "vyapaarclaw"
version = "0.1.0"
description = "VyapaarClaw Engine — AI Financial Governance MCP Server powering the VyapaarClaw OpenClaw Framework"
readme = "README.md"
authors = [
    { name = "guglxni", email = "aaryanguglani.cs21@rvce.edu.in" }
]
license = "AGPL-3.0-only"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.0.0",
    "razorpay>=1.4.0",
    "httpx>=0.27.0",
    "redis[hiredis]>=5.0.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "openai>=1.0.0",
    "textual>=8.1.1",
    "rich>=14.3.2",
    "holidays>=0.65",
    "forex-python>=1.8",
    "numpy-financial>=1.0.0",
    "fpdf2>=2.8.0",
    "networkx>=3.2",
    "transitions>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "fakeredis[lua]>=2.21.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "types-redis>=4.6.0",
]

[project.scripts]
vyapaarclaw = "vyapaar_mcp:main"
vyapaarclaw-tui = "vyapaar_mcp.tui.app:run_tui"

[build-system]
requires = ["uv_build>=0.8.11,<0.11.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "vyapaar_mcp"

[tool.mypy]
strict = true
strict_optional = false
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
warn_unused_ignores = false
disallow_untyped_decorators = false
disallow_untyped_calls = false
disallow_subclassing_any = false
disallow_any_generics = false

[[tool.mypy.overrides]]
module = [
    "textual.*",
    "sklearn.*",
    "asyncpg.*",
    "razorpay.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = [
    "RUF001",  # Ambiguous unicode chars in string literals (intentional info icons)
    "RUF002",  # Ambiguous unicode in docstrings (en-dash in prose)
    "E741",    # Ambiguous variable name (used in score calculations)
]

[tool.ruff.lint.per-file-ignores]
"src/vyapaar_mcp/reputation/anomaly.py" = ["N806"]
"src/vyapaar_mcp/db/postgres.py" = ["S608"]  # SQL uses parameterized $N placeholders, not string interpolation
"src/vyapaar_mcp/config.py" = ["S104"]  # 0.0.0.0 binding is intentional for container deployments
"src/vyapaar_mcp/server.py" = ["S104"]
